Service fabric (Early Rreview Swagger DO NOT GENERATE ANY SDKS)
This swagger will be checked in only for documentation preview. the
documnetation team currently doesnt offer a sandbox or local tools for
swagger authors to generate and preview there documentation.

Service fabric team need to get a preview to there swagger and the
documtatio team can only generate it through master